optimization api code
