+++ lib/credentials.js
