--- lib/database.js
@@ -1,5 +1,0 @@
-var tours = [
-	{ id: 0, name: 'Hood River', price: 99.99 },
-	{ id: 1, name: 'Oregon Coast', price: 149.95 },
-];
-exports.tours = tours
