+++ views/register.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
 <head>
-	<title>apitest</title>
+	<title>通过邮箱发送验证码注册账号</title>
 	<script src="http://code.jquery.com/jquery-2.0.2.min.js"></script>
 </head>
 <body>
@@ -21,7 +21,7 @@
 				var reg = /^([a-zA-Z]|[0-9])(\w|\-)+@[a-zA-Z0-9]+\.([a-zA-Z]{2,4})$/;
 				if(reg.test(email)){
 					$.ajax({
-						url:'http://192.168.1.252:3000/api/emails',//地址
+						url:'http://192.168.1.3:3000/api/emails',//地址
 						dataType:'json',//数据类型
 						type:'POST',//类型
 						timeout:2000,//超时
@@ -54,7 +54,7 @@
 				var _form = $('form[name=form]').serializeObject();
 				console.log(_form)
 				$.ajax({
-					url:'http://192.168.1.252:3000/api/register',//地址
+					url:'http://192.168.1.3:3000/api/register',//地址
 					dataType:'json',//数据类型
 					type:'POST',//类型
 					timeout:2000,//超时
