+++ views/testLocalDate.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
 <head>
-	<title>apitest</title>
+	<title>测试接口返回本地假数据</title>
 	<script src="http://code.jquery.com/jquery-2.0.2.min.js"></script>
 </head>
 <body>
@@ -10,7 +10,7 @@
 		$(function(){
 			$('button[name=testAPI]').click(function(){
 				$.ajax({
-					url:'http://192.168.1.3:3000/api/tours',//地址
+					url:'http://192.168.1.3:3000/api/loaclDate',//地址
 					dataType:'json',//数据类型
 					type:'GET',//类型
 					timeout:2000,//超时
