+++ views/uploadLocal.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
 <head>
-	<title>apitest</title>
+	<title>上传文件存储本地</title>
 	<script src="http://code.jquery.com/jquery-2.0.2.min.js"></script>
 </head>
 <body>
